Move to a translate util function.
