+++ lib/elements/autocomplete.js
@@ -3,7 +3,7 @@
 const color = require('kleur');
 const Prompt = require('./prompt');
 const { cursor } = require('sisteransi');
-const { style, clear, figures, strip } = require('../util');
+const { style, clear, figures, strip, translate } = require('../util');
 
 const getVal = (arr, i) => arr[i] && (arr[i].value || arr[i].title || arr[i]);
 const getTitle = (arr, i) => arr[i] && (arr[i].title || arr[i].value || arr[i]);
@@ -32,7 +32,7 @@ class AutocompletePrompt extends Prompt {
     this.select = opts.initial || opts.cursor || 0;
     this.fallback = opts.fallback || opts.initial !== void 0 ?
       `${figures.pointerSmall} ${getTitle(this.choices, this.initial)}` :
-      `${figures.pointerSmall} ${this.translate('noMatchesFound')}`;
+      `${figures.pointerSmall} ${translate('noMatchesFound')}`;
     this.suggestions = [];
     this.input = '';
     this.limit = opts.limit || 10;
