+++ lib/elements/confirm.js
@@ -1,6 +1,6 @@
 const color = require('kleur');
 const Prompt = require('./prompt');
-const { style, clear } = require('../util');
+const { style, clear, translate } = require('../util');
 const { erase, cursor } = require('sisteransi');
 
 /**
@@ -66,10 +66,8 @@ class ConfirmPrompt extends Prompt {
           color.bold(this.msg),
           style.delimiter(this.done),
           this.done
-            ? this.translate(this.value ? 'yes' : 'no')
-            : color.gray(
-              this.translate(this.initialValue ? 'yesPrimaryOption' : 'noPrimaryOption')
-            )
+            ? translate(this.value ? 'yes' : 'no')
+            : color.gray(translate(this.initialValue ? 'yesPrimaryOption' : 'noPrimaryOption'))
         ].join(' ')
     );
   }
