+++ lib/elements/multiselect.js
@@ -3,7 +3,7 @@
 const color = require('kleur');
 const Prompt = require('./prompt');
 const { cursor } = require('sisteransi');
-const { clear, figures, style } = require('../util');
+const { clear, figures, style, translate } = require('../util');
 
 /**
  * MultiselectPrompt Base Element
@@ -21,7 +21,7 @@ class MultiselectPrompt extends Prompt {
     super(opts);
     this.msg = opts.message;
     this.cursor = opts.cursor || 0;
-    this.hint = opts.hint || this.translate('hintMultiselect');
+    this.hint = opts.hint || translate('hintMultiselect');
     this.maxChoices = opts.max;
     this.value = opts.choices.map(v => Object.assign({ title: v.value, selected: false }, v));
     this.clear = clear('');
