+++ lib/elements/number.js
@@ -1,7 +1,7 @@
 const color = require('kleur');
 const Prompt = require('./prompt');
 const { cursor, erase } = require('sisteransi');
-const { style, clear, figures, strip } = require('../util');
+const { style, clear, figures, strip, translate } = require('../util');
 
 const isNumber = /[0-9]/;
 const isValidChar = /\.|-/;
@@ -37,7 +37,7 @@ class NumberPrompt extends Prompt {
     this.inc = opts.increment || 1;
     this.min = isDef(opts.min) ? opts.min : -Infinity;
     this.max = isDef(opts.max) ? opts.max : Infinity;
-    this.errorMsg = this.translate('errorInvalid');
+    this.errorMsg = translate('errorInvalid');
     this.validator = opts.validate || (() => true);
     this.color = `cyan`;
     this.value = ``;
