+++ lib/elements/prompt.js
@@ -16,7 +16,6 @@ class Prompt extends EventEmitter {
 
     this.in = opts.in || process.stdin;
     this.out = opts.out || process.stdout;
-    this.messages = opts.messages || {};
 
     const rl = readline.createInterface(this.in);
     readline.emitKeypressEvents(this.in, rl);
@@ -55,10 +54,6 @@ class Prompt extends EventEmitter {
 
   bell() {
     this.out.write(beep);
-  }
-
-  translate(key) {
-    return this.messages[key];
   }
 }
