+++ lib/elements/select.js
@@ -2,7 +2,7 @@
 
 const color = require('kleur');
 const Prompt = require('./prompt');
-const { style, clear, figures } = require('../util');
+const { style, clear, figures, translate } = require('../util');
 const { erase, cursor } = require('sisteransi');
 
 /**
@@ -19,7 +19,7 @@ class SelectPrompt extends Prompt {
   constructor(opts={}) {
     super(opts);
     this.msg = opts.message;
-    this.hint = opts.hint || this.translate('hintSelect');
+    this.hint = opts.hint || translate('hintSelect');
     this.cursor = opts.initial || 0;
     this.values = opts.choices || [];
     this.value = opts.choices[this.cursor].value;
