+++ lib/elements/text.js
@@ -1,7 +1,7 @@
 const color = require('kleur');
 const Prompt = require('./prompt');
 const { cursor } = require('sisteransi');
-const { style, clear, strip, figures } = require('../util');
+const { style, clear, strip, figures, translate } = require('../util');
 
 /**
  * TextPrompt Base Element
@@ -22,7 +22,7 @@ class TextPrompt extends Prompt {
     this.initial = opts.initial || ``;
     this.validator = opts.validate || (() => true);
     this.value = ``;
-    this.errorMsg = this.translate('errorInvalid');
+    this.errorMsg = translate('errorInvalid');
     this.cursor = Number(!!this.initial);
     this.clear = clear(``);
     this.render();
