+++ lib/index.js
@@ -2,21 +2,10 @@
 
 const util = require('util');
 const prompts = require('./prompts');
+const { translate } = require('./util');
 
 const passOn = ['suggest', 'format', 'onState', 'validate'];
 const noop = () => {};
-const messages = {
-  errorInvalid: 'Please enter a valid value',
-  hintMultiselect: '- Space to select. Return to submit.',
-  hintSelect: '- Use arrow-keys. Return to submit.',
-  messageRequired: 'Prompt message is required',
-  no: 'no',
-  noPrimaryOption: '(y/N)',
-  noMatchesFound: 'No matches found',
-  typeUndefined: 'Prompt type (%s) is not defined',
-  yes: 'yes',
-  yesPrimaryOption: '(Y/n)',
-};
 
 /**
  * Prompt for a series of questions
@@ -46,7 +35,7 @@ async function prompt(questions=[], { onSubmit=noop, onCancel=noop }={}) {
     }
 
     if (typeof question.message !== 'string') {
-      throw new Error(messages.messageRequired);
+      throw new Error(translate('messageRequired'));
     }
 
     // update vars in case they changed
@@ -56,11 +45,8 @@ async function prompt(questions=[], { onSubmit=noop, onCancel=noop }={}) {
     if (!type) continue;
 
     if (prompts[type] === void 0) {
-      throw new Error(util.format(messages.typeUndefined, type));
+      throw new Error(util.format(translate('typeUndefined'), type));
     }
-
-    // pass translated messages
-    question.messages = messages;
 
     try {
       answer = await prompts[type](question);
@@ -83,8 +69,8 @@ function inject(obj) {
   }
 }
 
-function translate(customMessages) {
-  Object.assign(messages, customMessages);
+function localize(customMessages) {
+  translate.localize(customMessages);
 }
 
-module.exports = Object.assign(prompt, { prompt, prompts, inject, translate });
+module.exports = Object.assign(prompt, { prompt, prompts, inject, localize });
