+++ lib/util/index.js
@@ -5,5 +5,6 @@ module.exports = {
   clear: require('./clear'),
   style: require('./style'),
   strip: require('./strip'),
-  figures: require('./figures')
+  figures: require('./figures'),
+  translate: require('./translate'),
 };
